code cleanup + style seperation
